Remove events.go
