--- events.go
@@ -1,32 +1,0 @@
-package empire
-
-import (
-	"encoding/json"
-	"fmt"
-	"io"
-
-	"github.com/docker/docker/pkg/jsonmessage"
-	"github.com/remind101/empire/pkg/image"
-)
-
-// FakeDockerPull returns a slice of events that look like a docker pull.
-func FakeDockerPull(img image.Image, w io.Writer) error {
-	messages := []jsonmessage.JSONMessage{
-		{Status: fmt.Sprintf("Pulling repository %s", img.Repository)},
-		{Status: fmt.Sprintf("Pulling image (%s) from %s", img.Tag, img.Repository), Progress: &jsonmessage.JSONProgress{}, ID: "345c7524bc96"},
-		{Status: fmt.Sprintf("Pulling image (%s) from %s, endpoint: https://registry-1.docker.io/v1/", img.Tag, img.Repository), Progress: &jsonmessage.JSONProgress{}, ID: "345c7524bc96"},
-		{Status: "Pulling dependent layers", Progress: &jsonmessage.JSONProgress{}, ID: "345c7524bc96"},
-		{Status: "Download complete", Progress: &jsonmessage.JSONProgress{}, ID: "a1dd7097a8e8"},
-		{Status: fmt.Sprintf("Status: Image is up to date for %s", img)},
-	}
-
-	enc := json.NewEncoder(w)
-
-	for _, m := range messages {
-		if err := enc.Encode(&m); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
