+++ resolver.go
@@ -1,14 +1,16 @@
 package empire
 
 import (
+	"encoding/json"
+	"fmt"
 	"io"
 
 	"golang.org/x/net/context"
 
-	"github.com/remind101/empire/pkg/image"
-
+	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/fsouza/go-dockerclient"
 	"github.com/remind101/empire/pkg/dockerutil"
+	"github.com/remind101/empire/pkg/image"
 )
 
 type Resolver interface {
@@ -63,4 +65,26 @@ func (r *dockerResolver) pullImage(ctx context.Context, img image.Image, output 
 		OutputStream:  output,
 		RawJSONStream: true,
 	})
+}
+
+// FakeDockerPull returns a slice of events that look like a docker pull.
+func FakeDockerPull(img image.Image, w io.Writer) error {
+	messages := []jsonmessage.JSONMessage{
+		{Status: fmt.Sprintf("Pulling repository %s", img.Repository)},
+		{Status: fmt.Sprintf("Pulling image (%s) from %s", img.Tag, img.Repository), Progress: &jsonmessage.JSONProgress{}, ID: "345c7524bc96"},
+		{Status: fmt.Sprintf("Pulling image (%s) from %s, endpoint: https://registry-1.docker.io/v1/", img.Tag, img.Repository), Progress: &jsonmessage.JSONProgress{}, ID: "345c7524bc96"},
+		{Status: "Pulling dependent layers", Progress: &jsonmessage.JSONProgress{}, ID: "345c7524bc96"},
+		{Status: "Download complete", Progress: &jsonmessage.JSONProgress{}, ID: "a1dd7097a8e8"},
+		{Status: fmt.Sprintf("Status: Image is up to date for %s", img)},
+	}
+
+	enc := json.NewEncoder(w)
+
+	for _, m := range messages {
+		if err := enc.Encode(&m); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
